requirements.txt
Django==4.2.11
djangorestframework==3.14.0
psycopg2-binary==2.9.9
django-allauth==0.57.0
django-crispy-forms==2.3
crispy-bootstrap5==2026.3
cloudinary==1.36.0
django-cloudinary-storage==0.3.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
requests==2.31.0
dj-database-url==2.1.0
